requirements.txt
openai
python-dotenv
requests
numpy
pymupdf
pypdf
sqlalchemy
fastapi
uvicorn
